* Fix missing long -> int32_t change from an old patch by Maximiliano Curia.

+++ abuse/trunk/src/include/compiled.hpp
@@ -1,7 +1,7 @@
 #ifndef __COMPILED_HPP_
 #define __COMPILED_HPP_
 #include "macs.hpp"
-extern long S_fall_start,S_falling,S_landing,S_pounce_wait,
+extern int32_t S_fall_start,S_falling,S_landing,S_pounce_wait,
             S_turn_around,S_fire_wait,S_ceil_fire,S_top_walk,
             S_blown_back_dead,S_jump_up,S_hiding,S_weapon_fire,
 	    S_hanging,S_blocking,S_rotate,S_climbing,S_climb_off,
